fix(transport): #1278 review consensus — S1+S3+S4+S5+S6 per consensus panel
Panel (architect, security-auditor, backend-dev) consensus on 7 remaining
findings — apply all except S2 (deferred to follow-up). See
artifacts/analyses/1278-fix-cycle-consensus.mdx.

S1 — WorkerPoolClient.stop() now awaits the heartbeat task (catching
CancelledError) before unsubscribing the subscription. Closes the race
where _on_heartbeat could fire after unsubscribe() began.

S3 + B5(image) — ImageGenParams moved from nats_image_client.py to
nats_image_codec.py (codec owns the wire-shape dataclass; previous
codec→client import inverted the dependency direction). nats_image_client
re-exports for backward compat. Image client drops 97L → 79L (under
spec sub-limit).

S4 — Removed `assert isinstance(result, Ok)` at 4 codec sites (llm,
tts, stt, image). Pyright already narrows after Err branch; asserts are
stripped under `python -O` violating the "never raises" codec contract.
Drops `Ok` from codec imports.

S5 — WorkerPoolClient._sub now typed as `Subscription | None` (under
TYPE_CHECKING) instead of `Any`. Restores type safety on .unsubscribe().

S6 — folder_exemptions.txt rationale comment for src/lyra/nats updated to
cite both deletions (_worker_client_base.py + nats_llm_client.py) and the
codec additions. Count corrected 18 → 16 to match actual .py file count.

S2 deferred: lyra.transport not in importlinter layers — requires moving
worker_registry out of lyra.nats (cross-Phase-1 scope). DEBT-tracked.

Bonus: SanitizedError.detail has no construction-time validator (preventive
follow-up, not a current vulnerability) — surfaced by security-auditor in
consensus, deferred to separate issue.
